+++ src/test/java/net/pwall/json/TestLong.java
@@ -41,31 +41,31 @@ class TestLong {
         JSONValue value = JSON.parse("123456789012");
         assertTrue(value instanceof JSONLong);
         assertEquals("123456789012", value.toString());
-        assertEquals(123456789012L, ((JSONLong)value).get());
+        assertEquals(123456789012L, ((JSONLong)value).getValue());
         assertEquals("123456789012", value.toJSON());
 
         value = JSON.parse("-123456789012");
         assertTrue(value instanceof JSONLong);
         assertEquals("-123456789012", value.toString());
-        assertEquals(-123456789012L, ((JSONLong)value).get());
+        assertEquals(-123456789012L, ((JSONLong)value).getValue());
         assertEquals("-123456789012", value.toJSON());
 
         value = JSON.parse("-1000000000000");
         assertTrue(value instanceof JSONLong);
         assertEquals("-1000000000000", value.toString());
-        assertEquals(-1000000000000L, ((JSONLong)value).get());
+        assertEquals(-1000000000000L, ((JSONLong)value).getValue());
         assertEquals("-1000000000000", value.toJSON());
 
         value = JSON.parse("9223372036854775807");
         assertTrue(value instanceof JSONLong);
         assertEquals("9223372036854775807", value.toString());
-        assertEquals(9223372036854775807L, ((JSONLong)value).get());
+        assertEquals(9223372036854775807L, ((JSONLong)value).getValue());
         assertEquals("9223372036854775807", value.toJSON());
 
         value = JSON.parse("-9223372036854775808");
         assertTrue(value instanceof JSONLong);
         assertEquals("-9223372036854775808", value.toString());
-        assertEquals(-9223372036854775808L, ((JSONLong)value).get());
+        assertEquals(-9223372036854775808L, ((JSONLong)value).getValue());
         assertEquals("-9223372036854775808", value.toJSON());
 
         value = JSON.parse("2147483647");
@@ -82,27 +82,27 @@ class TestLong {
     @Test
     void testConstructor1() {
         JSONLong long1 = new JSONLong(123);
-        assertEquals(123, long1.get());
+        assertEquals(123, long1.getValue());
         assertEquals("123", long1.toString());
         assertEquals("123", long1.toJSON());
 
         long1 = new JSONLong(0);
-        assertEquals(0, long1.get());
+        assertEquals(0, long1.getValue());
         assertEquals("0", long1.toString());
         assertEquals("0", long1.toJSON());
 
         long1 = new JSONLong(-1000);
-        assertEquals(-1000, long1.get());
+        assertEquals(-1000, long1.getValue());
         assertEquals("-1000", long1.toString());
         assertEquals("-1000", long1.toJSON());
 
         long1 = new JSONLong(Long.MAX_VALUE);
-        assertEquals(Long.MAX_VALUE, long1.get());
+        assertEquals(Long.MAX_VALUE, long1.getValue());
         assertEquals("9223372036854775807", long1.toString());
         assertEquals("9223372036854775807", long1.toJSON());
 
         long1 = new JSONLong(Long.MIN_VALUE);
-        assertEquals(Long.MIN_VALUE, long1.get());
+        assertEquals(Long.MIN_VALUE, long1.getValue());
         assertEquals("-9223372036854775808", long1.toString());
         assertEquals("-9223372036854775808", long1.toJSON());
 
@@ -130,32 +130,32 @@ class TestLong {
     @Test
     void testValueOf() {
         JSONLong long1 = JSONLong.valueOf(123456789012L);
-        assertEquals(123456789012L, long1.get());
+        assertEquals(123456789012L, long1.getValue());
         assertEquals("123456789012", long1.toString());
         assertEquals("123456789012", long1.toJSON());
 
         long1 = JSONLong.valueOf(0);
-        assertEquals(0L, long1.get());
+        assertEquals(0L, long1.getValue());
         assertEquals("0", long1.toString());
         assertEquals("0", long1.toJSON());
 
         long1 = JSONLong.valueOf(-1000);
-        assertEquals(-1000L, long1.get());
+        assertEquals(-1000L, long1.getValue());
         assertEquals("-1000", long1.toString());
         assertEquals("-1000", long1.toJSON());
 
         long1 = JSONLong.valueOf(Long.MAX_VALUE);
-        assertEquals(Long.MAX_VALUE, long1.get());
+        assertEquals(Long.MAX_VALUE, long1.getValue());
         assertEquals("9223372036854775807", long1.toString());
         assertEquals("9223372036854775807", long1.toJSON());
 
         long1 = JSONLong.valueOf(Long.MIN_VALUE);
-        assertEquals(Long.MIN_VALUE, long1.get());
+        assertEquals(Long.MIN_VALUE, long1.getValue());
         assertEquals("-9223372036854775808", long1.toString());
         assertEquals("-9223372036854775808", long1.toJSON());
 
         long1 = JSONLong.valueOf("123");
-        assertEquals(123L, long1.get());
+        assertEquals(123L, long1.getValue());
         assertEquals("123", long1.toString());
         assertEquals("123", long1.toJSON());